+++ webrtc/modules/audio_coding/codecs/ilbc/constants.h
@@ -81,7 +81,7 @@ extern const int16_t WebRtcIlbcfix_kAlpha[];
 
 extern const int16_t WebRtcIlbcfix_kEnhPolyPhaser[ENH_UPS0][ENH_FLO_MULT2_PLUS1];
 extern const int16_t WebRtcIlbcfix_kEnhWt[];
-extern const int16_t WebRtcIlbcfix_kEnhPlocs[];
+extern const size_t WebRtcIlbcfix_kEnhPlocs[];
 
 /* PLC tables */
