+++ webrtc/modules/audio_coding/neteq/expand.cc
@@ -467,8 +467,7 @@ void Expand::AnalyzeSignal(int16_t* random_vector) {
         correlation_length, correlation_lags, correlation_scale, -1);
 
     // Find maximizing index.
-    best_index = static_cast<size_t>(
-        WebRtcSpl_MaxIndexW32(correlation_vector2, correlation_lags));
+    best_index = WebRtcSpl_MaxIndexW32(correlation_vector2, correlation_lags);
     int32_t max_correlation = correlation_vector2[best_index];
     // Compensate index with start offset.
     best_index = best_index + start_index;
